requirements.txt
asgiref==3.5.0
Babel==2.10.1
bcrypt==3.2.0
cffi==1.15.0
Django==4.0.4
gunicorn==20.1.0
Pillow==9.1.0
psycopg2==2.9.3
pycparser==2.21
python-decouple==3.6
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
